+++ examples/rancher-workload-cluster-flexbot/terraform.tf
@@ -2,12 +2,12 @@ terraform {
   required_providers {
     flexbot = {
       source  = "igor-feoktistov/flexbot"
-      version = ">=1.6.6"
+      version = ">=1.8.0"
     }
     rancher2 = {
       source = "rancher/rancher2"
-      version = ">= 1.11.0"
+      version = ">=1.24.1"
     }
   }
-  required_version = ">= 0.13"
+  required_version = ">=0.15"
 }