+++ examples/rancher-workload-cluster-flexbot/variables.tf
@@ -6,6 +6,7 @@ variable "nodes" {
         model = string
         total_memory = string
       })
+      ip = string
       powerstate = string
       os_image = string
       seed_template = string
@@ -15,10 +16,24 @@ variable "nodes" {
         restore = bool
         snapshot_name = string
       })
+      maintenance = object({
+        execute = bool
+        synchronized_run = bool
+        wait_for_node_timeout = number
+        node_grace_timeout = number
+        tasks = list(string)
+      })
       snapshots = list(object({
         name = string
         fsfreeze = bool
       }))
+      labels = map(any)
+      taints = list(object({
+        key = string
+        value = string
+        effect = string
+      }))
+      force_update = bool
     }))
     workers = map(object({
       blade_spec = object({
@@ -26,6 +41,7 @@ variable "nodes" {
         model = string
         total_memory = string
       })
+      ip = string
       powerstate = string
       os_image = string
       seed_template = string
@@ -35,10 +51,24 @@ variable "nodes" {
         restore = bool
         snapshot_name = string
       })
+      maintenance = object({
+        execute = bool
+        synchronized_run = bool
+        wait_for_node_timeout = number
+        node_grace_timeout = number
+        tasks = list(string)
+      })
       snapshots = list(object({
         name = string
         fsfreeze = bool
       }))
+      labels = map(any)
+      taints = list(object({
+        key = string
+        value = string
+        effect = string
+      }))
+      force_update = bool
     }))
   })
 }
@@ -75,8 +105,8 @@ variable "node_config" {
       sp_org = string
       sp_template = string
       ssh_user = string
-      ssh_public_key_path = string
-      ssh_private_key_path = string
+      ssh_public_key = string
+      ssh_private_key = string
     })
     network = map(list(object({
       name = string
@@ -84,9 +114,12 @@ variable "node_config" {
       gateway = string
       dns_server1 = string
       dns_server2 = string
+      dns_server3 = string
       dns_domain = string
+      parameters = map(string)
     })))
     storage = object({
+      api_method = string
       zapi_version = string
     })
   })
@@ -97,6 +130,7 @@ variable "rancher_config" {
     api_url = string
     token_key = string
     cluster_name = string
+    cluster_api_server = string
     kubernetes_version = string
     s3_backup = object({
       region = string
@@ -110,7 +144,9 @@ variable "rancher_config" {
 }
 
 variable "pass_phrase" {
+  description = "Password to unlock credentials and keys"
   type = string
+  sensitive = true
 }
 
 variable "output_path" {